Consolidate Composites tests
